bispline_incidence {breedR}	R Documentation

Incidence Matrix of Bidimensional Splines

Description

Compute the incidence matrix as a tensor product of B-spline bases, given the knots, coordinates and order.

Usage

bispline_incidence(knots, xx, ord, sparse)

Arguments

knots
list with numeric vectors of knot positions with non-decreasing values for each dimension

xx
2-column matrix of coordinates at which to evaluate the bidimensional splines

ord
integer order of the spline functions (equals degree + 1)

sparse
logical indicating if the result should be given in sparse format

Details

Need at least 2*ord -1 knots (typically, 7) but in fact, we need at least 2*ord unless we set outer.ok = TRUE in splineDesign (which we do not want)

References

Eilers, P H C and B D Marx (2003). Multivariate calibration with temperature interaction using two-dimensional penalized signal regression. Chemometrics and Intelligent Laboratory Systems 66(2), 159-174.

[Package breedR version 0.12-5 Index]