breedR.option {breedR}	R Documentation

Set and get global options for breedR

Description

Set and get global options for breedR. The options are stored in the variable breedR.options in the .GlobalEnv-environment, and will therefore persist during the session. If you want to set some options permanently do it in a file names .breedRrc in your home directory. See Examples.

Usage

breedR.getOption(
  option = c("ar.eval", "breedR.bin", "splines.nok", "default.initial.variance",
    "col.seq", "col.div", "cygwin", "cygwin.home", "ssh.auth.sock", "remote.host",
    "remote.user", "remote.port", "remote.bin", "ssh.options")
)

breedR.setOption(...)

Arguments

option
The option to get. If missing or NULL, then breedR.getOption will display the current defaults, otherwise, option must be one of

ar.eval: numeric vector of values in (-1, 1) where the autoregressive parameters should be evaluated if not otherwise specified

splines.nok: a function of the number of individuals in a row which gives the number of knots (nok) to be used for a splines model, if not otherwise specified

default.initial.variance: a function of the numeric response vector or matrix which returns a default initial value for a variance component

col.seq: a vector with the specification of default extreme breedR col for sequential scales in spatial quantitative plots. See Details.

col.div: a vector with the specification of default extreme breedR col for diverging scales in spatial quantitative plots. See Details.

cygwin: the home of the Cygwin installation (default "C:/cygwin") [Remote computing for Windows only]

cygwin.home: the user's home in the Cygwin installation [Remote computing for Windows only]

ssh.auth.sock: the ssh bind-address (value of $SSH_AUTH_SOCK in the Cygwin-shell). [Remote computing for Windows only]

remote.host: (IP or DNS) address of a Linux server for remote computing

remote.user: user name with ssh-keys access to the server (see details)

breedR.bin: full path for breedR backend binaries

remote.port: port for ssh connection

remote.bin: path to the binaries directory in the remote installation of breedR. Usually the output of system.file('bin/linux', package='breedR') in a remote server's R-session.

ssh.options: ssh options. You shouldn't need to change this.

...
Option and value, like option=value or option, value; see the Examples.

Details

Sequential scales are used for variables not necessarily centered such as a response variable, or the fitted values of a model. The colour scale is built as a gradient between two extreme colours which are specified as hex codes or colour names in the option col.seq.

Diverging scales are used for plots such as residuals, centered (hopefully) around zero, with positive and negative values represented with different colours whose intensity is linked to the magnitude. The option col.div is a vector of two hex codes or colour names of the most intense colours.

Examples

## Set default values for the autoregressive parameters
breedR.setOption("ar.eval", 3*(-3:3)/10)
## alternative format
breedR.setOption(ar.eval = 3*(-3:3)/10) 
## check it 
breedR.getOption("ar.eval")

## Not run: 
# Set up some options permanently in $HOME/.breedRrc
writeLines(c("remote.host = '123.45.678.999'",
             "remote.user = 'uname'",
             "remote.bin  = 'remote/path/to/breedR/bin/linux'"),
           con = file.path(Sys.getenv('HOME'), '.breedRrc')

## End(Not run)

[Package breedR version 0.12-5 Index]