breedR.remote {breedR}	R Documentation

Perform a job remotely

Description

Assumes that all the relevant files are in the current directory.

Usage

breedR.remote(jobid, breedR.call, verbose = TRUE)

Arguments

jobid
character. A string uniquely identifying the current job.

breedR.call
character. A full string path to the executable program in the server.

verbose
logical. If TRUE (default) it shows informative messages.

[Package breedR version 0.12-5 Index]