breedr_ar {breedR}	R Documentation

Build an autoregressive model

Description

Given the coordinates of the observations, the autocorrelation parameters and the autofill logical value, computes the incidence matrix B and the covariance matrix U

Usage

breedr_ar(coordinates, rho, autofill = TRUE, ...)

Arguments

coordinates
matrix(-like) of observation coordinates

rho
numeric. Vector of length two with the autocorrelation parameter in each dimension, with values in the open interval (-1, 1).

autofill
logical. If TRUE (default) it will try to fill gaps in the rows or columns. Otherwise, it will treat gaps the same way as adjacent rows or columns.

...
Not used.

Value

a list with - the autocorrelation parameters - the coordinates original coordinates as a data frame - the incidence matrix (encoded as a vector) - the covariance matrix (of the full grid, in triplet form)

[Package breedR version 0.12-5 Index]