breedr_blocks {breedR}	R Documentation

Build a blocks model

Description

Given the coordinates of the observations, the *factor* identifying blocks, and the logical autofill, build the incidence and covariance matrices of a blocks model.

Usage

breedr_blocks(coordinates, id, autofill = TRUE, ...)

Arguments

coordinates
matrix(-like) of observation coordinates

id
factor of the same length as observations, giving the block id for each observation.

autofill
logical. If TRUE (default) it will try to fill gaps in the rows or columns. Otherwise, it will treat gaps the same way as adjacent rows or columns.

...
Not used.

[Package breedR version 0.12-5 Index]