breedr_effect {breedR}	R Documentation

Constructor for a generic breedR effect

Description

The breedr_effect-class is virtual. No object should be directly created with this constructor. This constructor is to be called from within non-virtual subclasses like generic, diagonal, spatial or additive_genetic.

Usage

breedr_effect(incidence)

## S3 method for class 'breedr_effect'
dim(x)

Arguments

incidence
matrix-like object

x
A breedr_effect.

Details

This constructor performs the arguments checks. But the implementation details (i.e., storage format and handling) is left for the subclasses.

Value

A list with a single element incidence.matrix.

[Package breedR version 0.12-5 Index]