build_grid {breedR}	R Documentation

Build an encompassing grid

Description

Build the minimal regularly-spaced grid containing a given set of points.

Usage

build_grid(coordinates, autofill = TRUE)

Arguments

coordinates
two-column matrix-like set of row and column coordinates of observational units

autofill
logical. If TRUE (default) it will try to fill missing rows or columns with missing observations. Otherwise, will treat individuals as neighbours even if they are across an empty line.

Details

Note that autofill = FALSE virtually removes the empty lines, considering the spacing as constant.

Value

The parameters defining the grid, and the index of the observed coordinates in the grid seen as a vector. More specifically,

origin

the coordinates of the first (with smallest row and column values) point in the grid

step

the separation between rows and columns

length

the number of points in each dimension

idx

the index of each observation in the vectorized grid

[Package breedR version 0.12-5 Index]