check_progsf90 {breedR}	R Documentation

Checks installation of PROGSF90 binaries

Description

Checks whether the binary dependencies are installed in the right directory. If not, allows calling the installer

Usage

check_progsf90(
  path = breedR.getOption("breedR.bin"),
  platform = breedR.os.type(),
  quiet = !interactive()
)

Arguments

path
directory to check for the presence of binaries. Default is defined in the package options, and it depends on the platform.

platform
Either "linux", "windows" or "mac". Default is current platform.

quiet
if TRUE, it won't ask whether to install missing binaries.

Details

This function does not check whether the binaries are for the right platform or architecture. It only checks the presence of files with the expected names.

[Package breedR version 0.12-5 Index]