check_var.ini {breedR}	R Documentation

Check initial variances specification

Description

If the user specified initial values, verify that all random effects were included. Otherwise, set default values. In any case, validate all initial values.

Usage

check_var.ini(x, random, response)

Arguments

x
list. user specification of var.ini (or NULL)

random
formula. user specification of random effects.

response
numeric vector or matrix.

Value

A list with initial covariance matrices for all random effects in the model. A logical attribute 'var.ini.default' is TRUE if values were set by default.

matrix of observation values.

[Package breedR version 0.12-5 Index]