compare.plots {breedR}	R Documentation

Compare two or more ggplots of the same kind

Description

This function presents several ggplots of the same type side by side under the same scale, while keeping annotations.

Usage

compare.plots(plots)

Arguments

plots
List of ggplots with meaningful names

Details

The names of the objects in the list will be used for facet labels.

[Package breedR version 0.12-5 Index]