competition {breedR}	R Documentation

Build a virtual competition model

Description

Given the coordinates of a set of observations, a decay parameter and a structure matrix, compute the incidence matrix of competition, and return a random effect with the given structure.

Usage

competition(coordinates, covariance, precision, decay, autofill = TRUE)

Arguments

coordinates
two-column matrix-like set of row and column coordinates of observational units

covariance
matrix-like object

precision
matrix-like object

decay
numeric. The positive value of the decay parameter α. Typically 1 or 2. See Details.

autofill
logical. If TRUE (default) it will try to fill missing rows or columns with missing observations. Otherwise, will treat individuals as neighbours even if they are across an empty line.

Details

The competition model attributes to each individual a random effect of competition with variance σ_{a_c}^2, which impacts the phenotype of the neighbours rather than its own phenotype.

Conversely, the effect of the competition over one's phenotype is given by the additive-genetic competition effects of the neighbours, weighted by the relative distances. If α is the decay parameter and a_c is the random competition effect of a neighbour at distance d, then the Weighted Neighbour Competition effect over one's phenotype is given by

∑_{neighbours} k (1/d)^α a_c,

where k is a normalizing constant which makes Var(wnc) = σ_{a_c}^2 and independent of the number of neighbours.

Value

An object inheriting from spatial.

[Package breedR version 0.12-5 Index]