determine.n.knots {breedR}	R Documentation

Determine a sensible number of knots

Description

This function computes a reasonable number of inner knots to be used for a basis of unidimensional B-splines

Usage

determine.n.knots(n, cutoff = 4, rate = 0.3)

Arguments

n
an integer vector of sample sizes

cutoff
a numeric vector of cutoff values

rate
a numeric vector of rates at which the number of knots increases with the sample size

Value

An integer vector with the number of knots for each sample size

References

Ruppert, D. (2002). Selecting the number of knots for penalized splines. Journal of Computational and Graphical Statistics 11, 735–757.

[Package breedR version 0.12-5 Index]