diagonal {breedR}	R Documentation

Constructor for a diagonal random effect

Description

Builds a breedR random effect with an index incidence matrix and a diagonal covariance matrix, with one level for each value of x.

Usage

diagonal(x)

Arguments

x
a numeric covariate or a factor.

Details

Uses sparse storage. It does not support nesting (yet). So it is not possible to build random regression coefficients for each level of a grouping factor. This is in the TODO list.

Value

A random effect with element incidence.matrix and a diagonal structure.matrix

[Package breedR version 0.12-5 Index]