distribute_knots_uniformgrid {breedR}	R Documentation

Distribute knots uniformly in a grid

Description

For each dimension, places the given number of knots evenly spaced covering the range of coordinates plus a small margin.

Usage

distribute_knots_uniformgrid(coordinates, n.knots, autofill)

Arguments

coordinates
matrix(-like) of observation coordinates

n.knots
numeric. Vector of length two with an integer number of knots in each dimension.

autofill
logical. If TRUE (default) it will try to fill gaps in the rows or columns. Otherwise, it will treat gaps the same way as adjacent rows or columns.

Details

The margin is calculated as half the median separation between observations. Furthermore, three more knots are added with equal spacing at each side, for each dimension.

[Package breedR version 0.12-5 Index]