douglas {breedR}	R Documentation

Multi-site Douglas-fir dataset

Description

Provenance test with measurements of height and circumference

Format

A dataframe with 1021 individuals and the following 9 variables

self id of the tree

dad id of sire or 0 if unknown

mum id of dam or 0 if unknown

orig factor origin of the seeds with 11 levels partially crossed with site

site factor site with 3 levels

block factor block with 127 levels nested within site

x, y coordinates (in m)

H02:H05heights as measured in 2002-2005

C13 circumference as measured in 2013

AN factor angle quality with 5 levels

BR factor branching quality with 5 levels

Details

The dataset includes measurements of height, circumference, angle and branching, for a progeny from 11 different origins which were planted in 1998 in 3 different sites.

Only 4 out of 11 origins are available in all the three sites.

Circumference was measured in all the three sites in 2013, but measurements of height were taken in 2002, 2003 and 2004 for site 3, and in 2005 for site 2. Measurements of angle and branching are only available for site 3.

The spacing between trees is different among sites, and in general between rows and columns. Coordinate units are meters.

Examples

data(douglas)

## Individuals by origin and site
with(douglas, table(orig, site))

## Measurements by variable and site
library(tidyr)
library(dplyr)
douglas %>%
 gather(variable, value, H02:BR, na.rm = TRUE) %>% 
 with(., table(variable, site))
 
## Visualise trials
library(ggplot2)
ggplot(douglas, aes(x, y)) +
  geom_point() + 
  facet_wrap(~ site)
 

[Package breedR version 0.12-5 Index]