effect_group {breedR}	R Documentation

Constructor for a group of effects

Description

Builds an effect_group from a list of breer_effect elements.

Usage

effect_group(x, cov.ini, ntraits)

## S3 method for class 'effect_group'
dim(x)

Arguments

x
list of breedr_effect elements

cov.ini
initial covariance matrix for the estimation algorithm

ntraits
number of traits in the model

Details

Temporarily, this takes the cov.ini argument and includes it in the object. In the future, the initial covariance matrix will be a matter of the inference engine, not inherent to the model.

The 'ntraits' is used to check the dimension of the initial variance matrix.

Value

A list of breedr_effect elements.

[Package breedR version 0.12-5 Index]