extract_block {breedR}	R Documentation

Extract a block of lines from a section of the REML log

Description

Extract a block of lines from a section of the REML log

Usage

extract_block(l, x)

Arguments

l
numeric. Line number where the block starts. Just after the section heading.

x
character vector. Lines of the log.

Value

character vector. Lines of text corresponding to numeric values under the section.

Examples

  test_log <- 
    c("REML log",
      "Some info",
      "Covariance matrix",
      "   1.23E2   4.56E-02   7.89E-03    ",
      "   1.23E2   4.56E-02   7.89E-03    "
      )
  breedR:::extract_block(4, test_log)

[Package breedR version 0.12-5 Index]