fixef {breedR}	R Documentation

Extract fixed-effects estimates

Description

Extract the fixed-effects estimates

Usage

## S3 method for class 'remlf90'
fixef(object, ...)

Arguments

object
any fitted model object from which fixed effects estimates can be extracted.

...
not used

Details

Extract the estimates of the fixed-effects parameters from a fitted model.

Value

a named list of dataframes with the estimated coefficients and their standard error

Examples

    res <- remlf90(phe_X ~ gg + bl, data = globulus)
    fixef(res)

[Package breedR version 0.12-5 Index]