generic {breedR}	R Documentation

Build a generic model

Description

Check conformity of arguments and return a generic object.

Usage

generic(incidence, covariance, precision)

Arguments

incidence
matrix-like object

covariance
matrix-like object

precision
matrix-like object

Details

A generic random effect stores the incidence and structure matrices in Matrix form, which tries to take advantage of sparsity, if it exists.

Value

A list with elements incidence.matrix, structure.matrix and structure.type, which is a string indicating either covariance or precision.

[Package breedR version 0.12-5 Index]