get_structure {breedR}	R Documentation

Covariance structure of a breedR component

Description

This generic function returns the covariance or precision matrix of a breedR random effect or a group of effects.

Usage

get_structure(x)

## S3 method for class 'breedR'
get_structure(x)

## S3 method for class 'effect_group'
get_structure(x)

## S3 method for class 'breedr_effect'
get_structure(x)

Arguments

x
A breedr_effect.

Details

For effect_groups, it returns the common structure of all the elements in the group.

Methods (by class)

breedR: Return the structure matrices of all structured random effects

effect_group: Check that all elements share the same structure and return it.

breedr_effect: Return the structure matrix with an attribute indicating its type.

[Package breedR version 0.12-5 Index]