install_progsf90 {breedR}	R Documentation

Install PROGSF90 binary dependencies

Description

Copy the binaries for the specified platform into a directory.

Usage

install_progsf90(
  url = breedr_progsf90_repo(),
  dest = system.file("bin", package = "breedR"),
  platform = breedR.os.type(),
  arch = breedR.os.32or64bit(),
  quiet = !interactive()
)

Arguments

url
where to download the files from

dest
destination directory for the binaries. Default is 'bin' under the current installation dir.

platform
what version of the binaries are to be installed. Default is current platform.

arch
Either "32" or "64". Coerced to string if necessary.

quiet
logical. Whether not to display messages.

Details

The url can be either of form http:// or of form file:// for local urls.

[Package breedR version 0.12-5 Index]