is_numericlog {breedR}	R Documentation

Test whether a string from a REML log is numeric

Description

A numeric string is a text string that contains only numbers that can be expressed in scientific format.

Usage

is_numericlog(x)

Arguments

x
character vector.

Details

Some alphabetic symbols are expected. For instance, in 1.23E-02. However, while spaces are expected, an empty line is not considered numeric.

Value

logical value.

Examples

breedR:::is_numericlog("  1.23E-02   1  ")
breedR:::is_numericlog("  var   1  ")
breedR:::is_numericlog("   ")

[Package breedR version 0.12-5 Index]