loc_grid {breedR}	R Documentation

Lattice of spatial locations

Description

Returns a list row and column coordinates of observations

Usage

loc_grid(coordinates, autofill)

Arguments

coordinates
two-column matrix-like set of row and column coordinates of observational units

autofill
logical. If TRUE (default) it will try to fill missing rows or columns with missing observations. Otherwise, will treat individuals as neighbours even if they are across an empty line.

Value

list of row and column coordinates of spatial units

This functions converts observational coordinates into spatial coordinates. While in a observational dataset there can possibly be many or missing observations at one single location, this function returns the list of row and column coordinates of a grid which contains all the observations

[Package breedR version 0.12-5 Index]