m1 {breedR}	R Documentation

A small Metagene synthesized dataset

Description

Simulated progeny of one single generation under phenotypic selection and spatial arrangement. The dataset includes the true Breeding values, the environmental effect and the observed phenotype for 1760 individuals.

Details

The generation of founders (generation 0) consist in 80 independent couples, each of which produces 20 descendants (10 of each sex).

The full dataset contains data for the 2*80 + 1600 = 1760 simulated individuals. However, the Metagene program does not simulate phenotypes for the founders.

The 1600 descendants were arranged at random in a 40 x 40 spatial grid

Examples

  # Load, summarize and visualize data
  data(m1)
  summary(m1)
  plot(m1)
  
  # Environmental component of the phenotype
  # (spatially structured)
  plot(m1, type = 'spatial')
  
  # The phenotypes are noisy observations around the 
  # true Breeding values with a standard deviation of about 7.3.
  library(ggplot2)
  qplot(BV_X, phe_X-BV_X, colour = dad, data = as.data.frame(m1)) + 
    geom_abline(intercept=0, slope=0, col='gray')

[Package breedR version 0.12-5 Index]