m4 {breedR}	R Documentation

Metagene synthesized dataset with four generations

Description

Simulated progeny of four generations under phenotypic selection and spatial arrangement. The dataset includes the true Breeding values, the environmental effect and the observed phenotype for the 6560 individuals.

Details

The generation of founders (generation 0) consist in 80 independent couples, each of which produces 20 descendants (10 of each sex). The second generation descend from 80 random couples of the best individuals among the 1600 members of the first generation. The same procedure is simulated to obtain the third and fourth generations.

The full dataset contains data for the 2*80 + 4*1600 = 6560 simulated individuals. However, the Metagene program does not simulate phenotypes for the founders.

The 6400 individuals from generations 1 to 4 were arranged at random in a 80 x 80 spatial grid

Examples

  # Load, summarize and visualize data
  data(m4)
  summary(m4)
  plot(m4)
  
  # Environmental component of the phenotype
  # (spatially structured)
  plot(m4, type = 'spatial')
  
  # The phenotypes are noisy observations around the Breeding values
  # with a standard deviation of about 10.
  library(ggplot2)
  qplot(BV_X, phe_X-BV_X, facets = .~gen, data = as.data.frame(m4)) + 
    geom_abline(intercept=0, slope=0, col='gray')

[Package breedR version 0.12-5 Index]