neighbours.at {breedR}	R Documentation

'move' an arrangement in a given direction

Description

'move' an arrangement in a given direction

Usage

neighbours.at(x, dir)

## S3 method for class 'matrix'
neighbours.at(x, dir)

## S3 method for class 'list'
neighbours.at(x, dir)

Arguments

x
matrix or list of matrices

dir
a direction in ('N', 'NE', 'E', 'SE', 'S', 'SW', 'W', 'NW')

Methods (by class)

matrix: Returns the matrix of neighbours in the specified direction

list: Returns the list of matrices of neighbours in the specified direction

[Package breedR version 0.12-5 Index]