parse.txtmat {breedR}	R Documentation

Parse a matrix from a text output robustly

Description

Each row of the matrix is a string. If rows are too long, they can continue in another line. Hence, the number of lines might be a multiple of the number of columns

Usage

parse.txtmat(x, names = NULL, square = TRUE)

Arguments

x
A character vector with space-separated numbers

names
A character vector with row and column names for the output matrix.

square
logical. Whether to assume that the matrix is square.

If the matrix is not necessarily

[Package breedR version 0.12-5 Index]