pf90_code_missing {breedR}	R Documentation

Determine a numeric code for missing observations

Description

This function returns a code outside the range of variation of the observed values.

Usage

pf90_code_missing(x)

Arguments

x
a numeric vector.

Details

If the range of variation strictly excludes zero, then it is used as the code for missing observations (which is the default code in PROGSF90). Otherwise, returns 1 - the second power of 10 greater than the maximum observed magnitude. This ensures a code an order of magnitude greater than the observed values. E.g., for observations in the range -40 – 28, the code is -999.

Examples

  breedR:::pf90_code_missing(rnorm(100))

[Package breedR version 0.12-5 Index]