pf90_default_heritability {breedR}	R Documentation

Default formula for heritability

Description

If all random effects are independent, and there is an additive-genetic effect, computes a default formula in PROGSF90 notation by dividing the genetic variance by the sum of all variance components plus the residual variance.

Usage

pf90_default_heritability(rglist, traits = NULL, quiet = FALSE)

Arguments

rglist
list of random groups in the parameters of a progsf90 object

traits
A character vector with trait names, or NULL for single trait.

quiet
logical. If FALSE, the function issues a message when it fails to build a formula.

Value

A character vector with one option specification per trait.

References

http://nce.ads.uga.edu/wiki/doku.php?id=readme.aireml#options

[Package breedR version 0.12-5 Index]