plot.remlf90 {breedR}	R Documentation

Spatial plot of a model's fit components

Description

Plots the predicted values of the component effects of the phenotype.

Usage

## S3 method for class 'remlf90'
plot(
  x,
  type = c("phenotype", "fitted", "spatial", "fullspatial", "residuals"),
  z = NULL,
  ...
)

Arguments

x
A breedR object

type
Character. Which component is to be represented in the map.

z
Optional. A numeric vector to be plotted with respect to the spatial coordinates. Overrides type.

...
Further layers passed to ggplot.

[Package breedR version 0.12-5 Index]