progsf90 {breedR}	R Documentation

progsf90 class

Description

This function parses a model frame and extracts the relevant fields that are to be written in the parameter, data and auxiliary files of the progsf90 programs.

Usage

progsf90(mf, weights, effects, opt = c("sol se"), res.var.ini = 10)

Arguments

mf
model.frame for fixed and diagonal random effects

weights
a vector of weights for the residual variance

effects
breedr_modelframe

opt
character. Options to be passed to Misztal's programs

res.var.ini
positive number. Initial value for the residual variance.

[Package breedR version 0.12-5 Index]