random {breedR}	R Documentation

Constructor for a random effect

Description

The random-class is virtual. No object should be directly created with this constructor. This constructor is to be called from within subclasses like generic, diagonal, spatial or additive_genetic.

Usage

random(incidence, covariance, precision)

Arguments

incidence
matrix-like object

covariance
matrix-like object

precision
matrix-like object

Details

This constructor performs the arguments and conformance checks. But the implementation details (i.e., storage format and handling) is left for the subclasses.

Value

A list with elements incidence.matrix, structure.matrix and structure.type, which is a string indicating either covariance or precision.

[Package breedR version 0.12-5 Index]