remote {breedR}	R Documentation

Control and view a remote breedR-queue of submitted jobs

Description

Control and view a remote breedR-queue of submitted jobs

Usage

breedR.qget(id, remove = TRUE)

breedR.qdel(id)

breedR.qstat(id)

breedR.qnuke()

breedR.remote_load(retry = 5)

Arguments

id
The job-id which is the output from breedR when the job is submitted, the job-number or job-name. For breedR.qstat, id is optional and if omitted all the jobs will be listed.

remove
Logical. If FALSE, leave the job on the server after retrieval, otherwise remove it (default).

retry
numeric. In case of connection failure, number of times to retry before giving up and returning NA.

Details

breedR.qstat shows job(s) on the server, breedR.qget fetches the results (and by default remove the files on the server), breedR.qdel removes a job on the server and breedR.qnuke removes all jobs on the server.

Finally, breedR.remote_load returns the current load in the server, as a percent. This should be used to check whether jobs can be safely submitted, and this is left to the user.

The recommended procedure is to use r <- remlf90(..., breedR.bin="submit") and then do r <- breedR.qget(r) at a later stage. If the job is not finished, then r will not be overwritten and this step can be repeated. The reason for this procedure, is that some information usually stored in the result object does not go through the remote server, hence have to be appended to the results that are retrieved from the server. Hence doing r <- remlf90(..., breedR.bin="submit") and then later retrieve it using r <- breedR.qget(1), say, then r does not contain all the required information.

Value

breedR.qstat returns an breedR.q-object with information about current jobs.

Remote computing under Windows

You need to install cygwin and ssh beforehand.

Setup

You need to configure the client and server machines so that passwordless SSH authentication works. See for example here

Furthermore, you need to configure breedR by setting the options remote.host, remote.user, remote.port and remote.bin. You can permanently set these options in the file .breedRrc in your home directory. See ?breedR.setOption.

See Also

remlf90

Examples

## Not run: 
r = remlf90(y~1, data = data.frame(y=rnorm(10)), breedR.bin = "submit")
summary(r)       # shows its status, same as breedR.qstat(r)
breedR.qstat()   # shows all jobs
r = breedR.qget(r, remove=FALSE)
breedR.qdel(1)
breedR.qnuke()
summary(r)   # results of the analysis

## End(Not run)

[Package breedR version 0.12-5 Index]