renderpf90 {breedR}	R Documentation

Render a progsf90 effect

Description

Translates breedR effects into progsf90 parameters and data.

Usage

renderpf90(x)

## Default S3 method:
renderpf90(x)

## S3 method for class 'fixed'
renderpf90(x)

## S3 method for class 'diagonal'
renderpf90(x)

## S3 method for class 'breedr_modelframe'
renderpf90(x, ntraits, weights)

## S3 method for class 'effect_group'
renderpf90(x)

## S3 method for class 'generic'
renderpf90(x)

## S3 method for class 'additive_genetic_animal'
renderpf90(x)

## S3 method for class 'additive_genetic_competition'
renderpf90(x)

## S3 method for class 'permanent_environmental_competition'
renderpf90(x)

## S3 method for class 'splines'
renderpf90(x)

## S3 method for class 'blocks'
renderpf90(x)

## S3 method for class 'ar'
renderpf90(x)

Arguments

x
object of class breedr_modelframe, effect_group or breedr_effect.

ntraits
integer. Number of traits in the model.

weights
logical. Whether there is an additional column of weights.

Details

This is an internal function. Not exported.

For the generic class, all matrices are converted to plain matrix-class, for exporting to files. The progsf90 model is either user_file or user_file_i depending on the type of structure matrix; i.e. respectively precision or covariance.

For the splines class, everything reduces to a generic effect with a covariance matrix

For the blocks class, everything reduces to a generic effect with a covariance matrix

For the ar class, everything reduces to a generic effect with a precision matrix

Value

The number of levels and type for each 'virtual' effect; the progsf90 model-name as appropriate; a file name and its content.

Methods (by class)

default: For unknown classes, just returns the object untouched with a warning.

fixed: For fixed effects.

diagonal: For diagonal effects. Assumed grouping variables.

breedr_modelframe: Render a full breedr_modelframe

effect_group: Render groups of effects into pf90 code

generic: Compute the parameters of a progsf90 representation of a generic effect.

additive_genetic_animal: Compute the parameters of a progsf90 representation of a additive_genetic_animal effect.

additive_genetic_competition: Compute the parameters of a progsf90 representation of a additive_genetic_competition effect.

permanent_environmental_competition: Compute the parameters of a progsf90 representation of a permanent_environmental_competition effect. Has the same incidence matrix than the additive_genetic_competition effect but an unstructured covariance matrix.

splines: Compute the parameters of a progsf90 representation of a splines effect.

blocks: Compute the parameters of a progsf90 representation of a blocks effect.

ar: Compute the parameters of a progsf90 representation of an AR effect.

See Also

Other renderpf90: renderpf90.matrix()

[Package breedR version 0.12-5 Index]