splat {breedR}	R Documentation

'Splat' arguments to a function

Description

Wraps a function in do.call, so instead of taking multiple arguments, it takes a single named list which will be interpreted as its arguments.

Usage

splat(flat)

Arguments

flat
function to splat

This is useful when you want to pass a function a row of data frame or array, and don't want to manually pull it apart in your function.

Borrowed from splat

Value

a function

Examples

  args <- replicate(3, runif(5), simplify = FALSE)
  identical(breedR:::splat(rbind)(args), do.call(rbind, args))

[Package breedR version 0.12-5 Index]