vgram.matrix {breedR}	R Documentation

Variogram of a matrix

Description

This function is slightly adapted from the homonym function in package fields. As such, the all the credit is for their authors.

Usage

vgram.matrix(dat, R = 5, dx = 1, dy = 1)

Arguments

dat
A matrix.

R
Maximum radius in distance units for computing the variogram.

dx
The spacing between rows of the matrix, in distance units.

dy
The spacing between columns of the matrix, in distance units.

Note

fields, Tools for spatial data Copyright 2004-2013, Institute for Mathematics Applied Geosciences University Corporation for Atmospheric Research Licensed under the GPL – www.gpl.org/licenses/gpl.html

[Package breedR version 0.12-5 Index]